+++ lib-image.go
@@ -127,7 +127,7 @@ func buildImageLib() {
 /*
     // set background colour, borders, etc with 'style'
     slhelp["graph"] = LibHelp{in: "handle,ox,oy,w,h[,attr]", out: "graph_handle", action: "Sets up a new graph."}
-    stdlib["graph"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["graph"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         return nil,nil
         if len(args)!=6 {
             return nil, errors.New("Bad arguments supplied to graph()")
@@ -150,7 +150,7 @@ func buildImageLib() {
     }
 
     slhelp["graph_output"] = LibHelp{in: "handle", out: "", action: "Sends the graph to the canvas."}
-    stdlib["graph_output"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["graph_output"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         return nil,nil
         if len(args)!=1 {
             return nil, errors.New("Bad arguments supplied to graph_output()")
@@ -177,7 +177,7 @@ func buildImageLib() {
 */
 
     slhelp["svg_circle"] = LibHelp{in: "handle,x,y,radius[,attributes]", out: "", action: "Draws a circle on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_circle"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_circle"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_circle()") }
         var handle string
         switch args[0].(type) {
@@ -207,7 +207,7 @@ func buildImageLib() {
 
     // Square(x int, y int, s int, style ...string)
     slhelp["svg_square"] = LibHelp{in: "handle,x,y,size[,attributes]", out: "", action: "Draws a square on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_square"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_square"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_square()") }
         var handle string
         switch args[0].(type) {
@@ -237,7 +237,7 @@ func buildImageLib() {
 
     // Grid(x int, y int, w int, h int, n int, s ...string)
     slhelp["svg_grid"] = LibHelp{in: "handle,x,y,w,h,n[,attributes]", out: "", action: "Draws a grid on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_grid"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_grid"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_grid()") }
         var handle string
         switch args[0].(type) {
@@ -269,7 +269,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_ellipse"] = LibHelp{in: "handle,x,y,w,h[,attributes]", out: "", action: "Draws an ellipse on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_ellipse"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_ellipse"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         // Ellipse(x int, y int, w int, h int, s ...string)
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_ellipse()") }
         var handle string
@@ -302,7 +302,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_plot"] = LibHelp{in: "handle,x,y[,attributes]", out: "", action: "Plots a point on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_plot"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_plot"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_plot()") }
         var handle string
         switch args[0].(type) {
@@ -332,7 +332,7 @@ func buildImageLib() {
 
     // Text(x int, y int, t string, s ...string)
     slhelp["svg_text"] = LibHelp{in: "handle,x,y,text[,attributes]", out: "", action: "Writes text to the SVG canvas [#i1]handle[#i0]."}
-    stdlib["svg_text"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_text"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_text()") }
         var handle string
         switch args[0].(type) {
@@ -364,7 +364,7 @@ func buildImageLib() {
 
     // Image(x int, y int, w int, h int, link string, s ...string)
     slhelp["svg_image"] = LibHelp{in: "handle,x,y,w,h,image_link[,attributes]", out: "", action: "Links an image to the SVG canvas [#i1]handle[#i0]."}
-    stdlib["svg_image"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_image"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_image()") }
         var handle string
         switch args[0].(type) {
@@ -400,7 +400,7 @@ func buildImageLib() {
 
 
     slhelp["svg_polygon"] = LibHelp{in: "handle,[]x,[]y[,attributes]", out: "", action: "Draws a polygon on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_polygon"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_polygon"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_polygon()") }
         var handle string
         switch args[0].(type) {
@@ -431,7 +431,7 @@ func buildImageLib() {
 
     // Polyline(x []int, y []int, s ...string)
     slhelp["svg_polyline"] = LibHelp{in: "handle,[]x,[]y[,attributes]", out: "", action: "Draws a polyline on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_polyline"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_polyline"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_polyline()") }
         var handle string
         switch args[0].(type) {
@@ -462,7 +462,7 @@ func buildImageLib() {
 
     // Rect(x int, y int, w int, h int, s ...string)
     slhelp["svg_rect"] = LibHelp{in: "handle,x,y,w,h[,attributes]", out: "", action: "Draws a rectangle on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_rect"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_rect"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_rect()") }
         var handle string
         switch args[0].(type) {
@@ -496,7 +496,7 @@ func buildImageLib() {
 
     // Roundrect(x int, y int, w int, h int, rx int, ry int, s ...string)
     slhelp["svg_roundrect"] = LibHelp{in: "handle,x,y,w,h,rx,ry[,attributes]", out: "", action: "Draws a rounded rectangle on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_roundrect"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_roundrect"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_roundrect()") }
         var handle string
         switch args[0].(type) {
@@ -534,7 +534,7 @@ func buildImageLib() {
 
 
     slhelp["svg_line"] = LibHelp{in: "handle,x,y,w,h[,attributes]", out: "", action: "Draws a line on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_line"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_line"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_line()") }
         var handle string
         switch args[0].(type) {
@@ -567,7 +567,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_title"] = LibHelp{in: "handle,title", out: "", action: "Places a title on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_title"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_title"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_title()") }
         var handle string
         switch args[0].(type) {
@@ -591,7 +591,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_desc"] = LibHelp{in: "handle,description", out: "", action: "Attach a description to the canvas [#i1]handle[#i0]."}
-    stdlib["svg_desc"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_desc"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments supplied to svg_desc()") }
         var handle string
         switch args[0].(type) {
@@ -616,7 +616,7 @@ func buildImageLib() {
 
 
     slhelp["svg_def"] = LibHelp{in: "handle", out: "", action: "Attach a definition to the canvas [#i1]handle[#i0]."}
-    stdlib["svg_def"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_def"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return nil, errors.New("Bad arguments supplied to svg_def()") }
         var handle string
         switch args[0].(type) {
@@ -630,7 +630,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_def_end"] = LibHelp{in: "handle", out: "", action: "Ends a definition in [#i1]handle[#i0]."}
-    stdlib["svg_def_end"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_def_end"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return nil, errors.New("Bad arguments supplied to svg_def_end()") }
         var handle string
         switch args[0].(type) {
@@ -644,7 +644,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_group"] = LibHelp{in: "handle[,attributes]", out: "", action: "Attach a group definition to the canvas [#i1]handle[#i0]."}
-    stdlib["svg_group"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_group"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 || len(args)>2 { return nil, errors.New("Bad arguments supplied to svg_group()") }
         var handle string
         switch args[0].(type) {
@@ -662,7 +662,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_group_end"] = LibHelp{in: "handle", out: "", action: "Ends a group definition in [#i1]handle[#i0]."}
-    stdlib["svg_group_end"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_group_end"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return nil, errors.New("Bad arguments supplied to svg_group_end()") }
         var handle string
         switch args[0].(type) {
@@ -677,7 +677,7 @@ func buildImageLib() {
 
     // Link(href string, title string)
     slhelp["svg_link"] = LibHelp{in: "handle,href,title", out: "", action: "Places a link on the canvas [#i1]handle[#i0]."}
-    stdlib["svg_link"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_link"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)==0 { return nil, errors.New("Bad arguments (count) supplied to svg_link()") }
         var handle string
         switch args[0].(type) {
@@ -704,7 +704,7 @@ func buildImageLib() {
 
     // LinkEnd()
     slhelp["svg_link_end"] = LibHelp{in: "handle", out: "", action: "Ends a link definition in [#i1]handle[#i0]."}
-    stdlib["svg_link_end"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_link_end"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 { return nil, errors.New("Bad arguments (count) supplied to svg_link_end()") }
         var handle string
         switch args[0].(type) {
@@ -723,7 +723,7 @@ func buildImageLib() {
 
 
     slhelp["svg_start"] = LibHelp{in: "filename,w,h[,attributes]", out: "svg_handle", action: "Returns a handle to a new SVG canvas. You must use this call to initiate the image creation process.\nReturns an empty string on failure or a handle name on success."}
-    stdlib["svg_start"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_start"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
 
         if len(args)<3 || len(args)>4 {
             return nil, errors.New("Bad arguments supplied to svg_start()")
@@ -746,7 +746,7 @@ func buildImageLib() {
     }
 
     slhelp["svg_end"] = LibHelp{in: "handle", out: "", action: "Signals completion of the SVG canvas [#i1]handle[#i0]."}
-    stdlib["svg_end"] = func(evalfs uint64,args ...interface{}) (ret interface{}, err error) {
+    stdlib["svg_end"] = func(evalfs uint32,args ...interface{}) (ret interface{}, err error) {
         if len(args)!=1 {
             return nil, errors.New("Bad arguments supplied to svg_end()")
         }
